backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
google-cloud-texttospeech
google-cloud-storage

google-adk>=1.17.0
numpy
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-gcp-trace
opentelemetry-exporter-gcp-logging
opentelemetry-exporter-gcp-monitoring
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-google-genai>=0.4b0
opentelemetry-instrumentation-vertexai>=2.0b0

google-auth
google-cloud-logging
